requirements.txt
# Парсинг и HTTP
requests>=2.31.0,<3
beautifulsoup4>=4.12.0,<5
lxml>=5.0.0,<6

# Загрузка страниц через браузер (обход 403)
playwright>=1.40.0,<2

# Генерация изображений
Pillow>=10.0.0,<11

# Опционально: генерация фото авто через нейросеть (DALL-E)
# Для использования: pip install openai && задать OPENAI_API_KEY
# openai>=1.0.0
